LayoutTests/media/video-controls-drag-expected.txt

This tests that videos with controls enabled are still draggable. In a browser, start the test by dragging the video outside its controls area.

EVENT(drag)
END OF TEST
